Update mask-sensitive-values.rst
I recently had to work on something related to this documentation page and I found the documentation lacking a bit. There was no specification about the type of values that mask_secret accepts, or the fact that it escapes any string passed, so it's not possible to pass patterns. 
Hopefully these changes will make it clearer.

I am also considering adding an overload for this method such that it will accept patterns for masking.